# scraper/api/openapi.yaml
openapi: 3.0.0
info:
  title: Prysm Scraper API
  version: 1.0.0
  description: API for the Prysm structure-aware web scraper
  contact:
    name: Pink Pixel
    url: https://pinkpixel.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:3000
    description: Development server

tags:
  - name: Jobs
    description: Scraping job operations
  - name: Results
    description: Scraping results operations

components:
  schemas:
    ScraperOptions:
      type: object
      properties:
        maxScrolls:
          type: integer
          description: Maximum number of scroll attempts
          default: 5
        scrollDelay:
          type: integer
          description: Delay in ms between scroll actions
          default: 2000
        waitForSelector:
          type: string
          description: CSS selector to wait for before starting
          nullable: true
        headless:
          type: boolean
          description: Whether to run Puppeteer in headless mode
          default: true
        proxy:
          type: string
          description: Proxy server URL
          nullable: true
        useResourceBlocker:
          type: boolean
          description: Whether to use resource blocker
          default: true
        bypassCloudflare:
          type: boolean
          description: Whether to attempt to bypass Cloudflare
          default: true
        handlePagination:
          type: boolean
          description: Whether to automatically handle pagination
          default: true
        paginationStrategy:
          type: string
          enum: [infinite, click, url]
          description: Force a specific pagination strategy
          nullable: true
        clickSelector:
          type: string
          description: CSS selector for click-based pagination
          nullable: true
    
    JobRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          description: URL to scrape
        options:
          $ref: '#/components/schemas/ScraperOptions'
        priority:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
          description: Job priority (1 = highest, 10 = lowest)
        webhook:
          type: string
          format: uri
          description: URL to receive webhook notifications when job completes
    
    JobResponse:
      type: object
      properties:
        jobId:
          type: string
          description: Unique identifier for the job
        status:
          type: string
          enum: [pending, processing, completed, failed, cancelled]
        url:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        error:
          type: string
          nullable: true
        progress:
          type: integer
          minimum: 0
          maximum: 100
    
    ScrapingResults:
      type: object
      properties:
        title:
          type: string
        content:
          type: array
          items:
            type: string
        metadata:
          type: object
        structureType:
          type: string
        paginationType:
          type: string
        extractionMethod:
          type: string
        url:
          type: string
          format: uri
    
    JobWithResults:
      allOf:
        - $ref: '#/components/schemas/JobResponse'
        - type: object
          properties:
            result:
              $ref: '#/components/schemas/ScrapingResults'
    
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object

  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    
    BadRequest:
      description: The request was malformed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    
    Unauthorized:
      description: Authentication is required or failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    
    TooManyRequests:
      description: Request rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
            
paths:
  /api/jobs:
    post:
      tags:
        - Jobs
      summary: Create a new scraping job
      description: Start a new web scraping job on the specified URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobRequest'
      responses:
        '202':
          description: Job accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    
    get:
      tags:
        - Jobs
      summary: List all jobs
      description: Get a list of all scraping jobs with optional filtering
      parameters:
        - name: status
          in: query
          description: Filter jobs by status
          schema:
            type: string
            enum: [pending, processing, completed, failed, cancelled]
        - name: limit
          in: query
          description: Maximum number of jobs to return
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          description: Offset for pagination
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/JobResponse'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
  
  /api/jobs/{jobId}:
    parameters:
      - name: jobId
        in: path
        required: true
        description: Job identifier
        schema:
          type: string
    
    get:
      tags:
        - Jobs
      summary: Get job status
      description: Get the current status and details of a specific job
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    
    delete:
      tags:
        - Jobs
      summary: Cancel or delete a job
      description: Cancel a running job or delete a completed job's data
      responses:
        '204':
          description: Job successfully cancelled or deleted
        '404':
          $ref: '#/components/responses/NotFound'
  
  /api/jobs/{jobId}/results:
    parameters:
      - name: jobId
        in: path
        required: true
        description: Job identifier
        schema:
          type: string
    
    get:
      tags:
        - Results
      summary: Get job results
      description: Retrieve the results of a completed scraping job
      responses:
        '200':
          description: Job results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobWithResults'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Job is not yet completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error' 